feat: add a use translation hook to allow better fallback to english
