+++ packages/notification-center/src/components/notification-center/components/notification-item/NotificationListItem.tsx
@@ -7,7 +7,7 @@ import { INovuTheme } from '../../../../store/novu-theme.context';
 import { useNovuThemeProvider } from '../../../../hooks/use-novu-theme-provider.hook';
 import { ActionContainer } from './ActionContainer';
 import { NotificationCenterContext } from '../../../../store/notification-center.context';
-import { I18NContext } from '../../../../store/i18n.context';
+import { useTranslations } from '../../../../hooks/use-translations';
 
 export function NotificationListItem({
   notification,
@@ -18,7 +18,7 @@ export function NotificationListItem({
 }) {
   const { theme: novuTheme } = useNovuThemeProvider();
   const { onActionClick, listItem } = useContext(NotificationCenterContext);
-  const { lang } = useContext(I18NContext);
+  const { lang } = useTranslations();
 
   function handleNotificationClick() {
     onClick(notification);
