+++ packages/notification-center/src/i18n/lang.ts
@@ -1,0 +1,26 @@
+import { EN } from './languages/en';
+
+export interface ITranslationContent {
+  readonly notifications: string;
+  readonly markAllAsRead: string;
+  readonly poweredBy: string;
+}
+
+export interface ITranslationEntry {
+  readonly translations: ITranslationContent;
+  readonly lang: string;
+}
+
+export const TRANSLATIONS: Record<I18NLanguage, ITranslationEntry> = {
+  en: EN,
+};
+
+/**
+ * Should use the short notations of the W3C internationalization document
+ * https://www.w3.org/International/O-charset-lang.html
+ *
+ * For example:
+ * - For English use "en"
+ * - For French use "fr"
+ */
+export type I18NLanguage = 'en';
