+++ packages/notification-center/src/i18n/languages/en.ts
@@ -1,0 +1,10 @@
+import { ITranslationEntry } from '../lang';
+
+export const EN: ITranslationEntry = {
+  translations: {
+    notifications: 'Notifications',
+    markAllAsRead: 'Mark all as read',
+    poweredBy: 'Powered By',
+  },
+  lang: 'en',
+};
