--- packages/notification-center/src/lang.ts
@@ -1,18 +1,0 @@
-export interface ITranslationEntry {
-  readonly translations: {
-    readonly notifications: string;
-    readonly markAllAsRead: string;
-    readonly poweredBy: string;
-  };
-  readonly lang: string;
-}
-
-export const TRANSLATIONS: Record<I18NLanguage, ITranslationEntry['translations']> = {
-  en: {
-    notifications: 'Notifications',
-    markAllAsRead: 'Mark all as read',
-    poweredBy: 'Powered By',
-  },
-};
-
-export type I18NLanguage = 'en';
